requirements.txt
requests==2.31.0
python-dotenv==0.19.0
tk==0.1.0
tqdm==4.66.1
pandas==1.5.3
openpyxl==3.1.2
python-docx>=0.8.11
PyYAML>=6.0.1
PyPDF2>=3.0.0
xlrd>=2.0.1 